Update copr container to run make on the Makefile if the srpm_path is specified
Description:
Currently, when attempting to package a project inside of a subdirectory of a repo the builder will attempt to run make
inside the root folder instead of the subdirectory. This patch specifies the `project_dirname` (if specified) when
building. Additionally, if the user specifies an srpm path it will first check to see if the srpm exists in the
specified path. If it doesn't exist, it will run make and assign the outdir to the environemnt variable. Once make
completes, it will upload the srpm to the copr repository where it will be built.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>