Auto increment release value when performing dev builds on copr
Description:

In development, we may need to push several builds to copr and then test them on multiple nodes. If a new build is
pushed to copr with the same version and release as the previous push, copr will build the rpm but it will not install
the newest package when attempting to install via yum. To resolve this issue, the builder will do the following:

1. Does an SRPM already exist? If so, just upload it.
2. Otherwise, is this a production build?
   yes:
      a. Build the SRPM using the provided spec file
      b. Upload the SRPM to copr
   no:
      a. Fetch the list of builds for the package being worked on
      b. Sort the list of builds by the "ended_on" timestamp such that the most recent build is first
      c. Extract the version and release from the most recent build
      d. Compare the version of the most recent build with the version in the spec file
         same:
            - Increment the release value from the last build by 1.
         different:
            - Set release to 1
      e. Updat the Release filed in the spec file to be: Release: <epoch>.<new_release_value>%{?dist}
      f. Build the SRPM from the new spec
      g. Upload the SRPM to copr

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>