Create parameter to customize make srpm call
In monorepos where multiple modules point to the top level,
we have a naming clash between make srpm tasks.

I suspect the correct way to handle this is to put each RPM level
component in a subdir, but in the interest of time, add a way to build
srpms with different task names.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>